FaceMaskDetection machine learning project
here we have used the HaarCascadeFace file for detection face . we've generated own dataset for this project by using python library Opencv and store that in form of an array by Numpy and visualized by Matplotlib library. after that  we used Support Vector Machine and Support vector Classifier from Scikit_learn library for train and test the data . after the trainning the data we used Principal Component Analysis for reducing dimensions as PCA, is a dimensionality-reduction method that is often used to reduce the dimensionality of large data sets, by transforming a large set of variables into a smaller one that still contains most of the information in the large set. finally, our model is ready for the action and streaming live performance.